Bugfix: yuv_color assumed to be a YUV color.
yuv_color made assumptions on the default zero-value of the underlying
colorspace. i.e. It assumed that the colorspace was YUV. (Dangerous assumtion)

The fix involved adding the assumtion that 0.0 is always the identity element
of the colorspace. A cleaner way to fix it would be to add a color-class that knows
which channels are present or not.