Approach abandoned, design needs rework.
This line of approach seems to be a dead end, but I'll leave this code here for
future reference if the issues uncovered are solved.

The current design attempts to allow (almost) all parts of a format template to
be a MiniCalc expression and defines a syntax to refer to all other parts of the
template.

This, however, breaks down in the presence of block_order. As each plane is padded to be
a multiple of the macro_block dimensions the definition of image_w and image_h becomes
undefined outside of a specific plane instance.

Several solutions have been considered, but have been rejected:

- Require the user to add the dependency themselves, (and check that the dependency
  has been added during parsing).
  Rejected because it breaks backwards compatability and makes block_order even
  harder to use, in my opinion.

- Translate image_w -> plane[i].image_w and disallow any use of image_w without a plane
  specifier outside of the plane blocks.
  Rejected because it is so complex (having to juggle N potentially different image dimensions),
  in the very least it will require some user testing to attempt this approach.